[jit] serialize all c++ frontend modules to a single CU.
Previously, every module would get its own CompilationUnit when saving
from the C++ frontend. That's bad because nothing checks that they have
uniquely qualified names or mangles them to make them unique.

This was okay when we were doing model.json, but once we serialize
modules like classes this will cause an error on import (when we try to
re-define the same class a bunch of times.

ghstack-source-id: ae0ed6be7f4ad528875a26b8546f2962ae8179e6
Pull Request resolved: https://github.com/pytorch/pytorch/pull/23645